+++ CHANGELOG.md
@@ -17,7 +17,7 @@ N/A
 - Creature: {Get|Set}WalkAnimation()
 
 ### Changed
-N/A
+- Events: `NWNX_ON_DM_SPAWN_OBJECT_*` now provides the resref as event data.
 
 ### Deprecated
 N/A